refactor(server): expose reusable server entrypoint
