+++ crates/izwi-server/Cargo.toml
@@ -40,6 +40,7 @@ base64 = { workspace = true }
 hound = { workspace = true }
 clap = { version = "4.5", features = ["derive", "env"] }
 async-trait = "0.1"
+izwi-hooks = { path = "../izwi-hooks" }
 izwi-agent = { path = "../izwi-agent" }
 
 config = { workspace = true }
